Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
pydantic-settings
passlib[argon2]
python-jose[cryptography]
email-validator
chromadb
pydantic[email]
sentence-transformers
openai
asyncpg
passlib
python-jose[cryptography]    
python-multipart
pymupdf
llama-index-core 
llama-index-readers-file 
llama-index-llms-ollama 
llama-index-embeddings-huggingface
groq 
websockets 
pyaudio 
SpeechRecognition
vapi-python
vapi_server_sdk